+++ ChangeLog.md
@@ -15,6 +15,7 @@ New Features:
 
 * The detailed report is no longer part of the table validator. See [issue #378](https://github.com/biocore/biom-format/issues/378).
 * `load_table` now accepts open file handles. See [issue #481](https://github.com/biocore/biom-format/issues/481).
+* `biom export-metadata` has been added to export metadata as TSV. See [issue #820](https://github.com/biocore/biom-format/issues/820).
 
 Bug fixes: